requirements.txt
aiogram==3.4.1
SQLAlchemy==2.0.20
aiosqlite==0.18.0
python-dotenv==1.0.0
alembic==1.11.1
aiofiles==23.2.1
aiohttp~=3.9.5
